rename project to match release package
