+++ JunimoFarm/JunimoFarm.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{FFA58F5D-536B-4D0A-BADA-04BABFEFAAB1}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TheJunimoExpress</RootNamespace>
-    <AssemblyName>TheJunimoExpress</AssemblyName>
+    <RootNamespace>JunimoFarm</RootNamespace>
+    <AssemblyName>JunimoFarm</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
