+++ JunimoFarm/Properties/AssemblyInfo.cs
